修复负载均衡时的bug.
