+++ txlcn-tracing/src/main/java/com/codingapi/txlcn/tracing/http/RestTemplateRequestInterceptor.java
@@ -45,8 +45,10 @@ public class RestTemplateRequestInterceptor implements ClientHttpRequestIntercep
     public ClientHttpResponse intercept(
             @NonNull HttpRequest httpRequest, @NonNull byte[] bytes,
             @NonNull ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
-        httpRequest.getHeaders().add(TracingConstants.HEADER_KEY_GROUP_ID, TracingContext.tracing().groupId());
-        httpRequest.getHeaders().add(TracingConstants.HEADER_KEY_APP_MAP, TracingContext.tracing().appMapBase64String());
+        if (TracingContext.tracing().hasGroup()) {
+            httpRequest.getHeaders().add(TracingConstants.HEADER_KEY_GROUP_ID, TracingContext.tracing().groupId());
+            httpRequest.getHeaders().add(TracingConstants.HEADER_KEY_APP_MAP, TracingContext.tracing().appMapBase64String());
+        }
         return clientHttpRequestExecution.execute(httpRequest, bytes);
     }
 }
