+++ txlcn-tracing/src/main/java/com/codingapi/txlcn/tracing/http/ribbon/DTXLoadbalancer.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2017-2019 CodingApi .
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.codingapi.txlcn.tracing.http.ribbon;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+
+/**
+ * Description: springcloud ribbon 开启注解配置
+ * Date: 19-02-07 上午 00:08
+ *
+ * @author lorne
+ */
+@EnableConfigurationProperties(DTXLoadbalancer.class)
+@ConfigurationProperties("tx-lcn.ribbon.loadbalancer.dtx")
+public class DTXLoadbalancer {
+
+    private boolean enabled;
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+}
