+++ txlcn-tracing/src/main/java/com/codingapi/txlcn/tracing/http/ribbon/LoadbalancerConfiguration.java
@@ -17,10 +17,7 @@ package com.codingapi.txlcn.tracing.http.ribbon;
 
 import com.netflix.loadbalancer.IRule;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.client.serviceregistry.Registration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -33,23 +30,11 @@ import org.springframework.context.annotation.Scope;
  * @author ujued
  */
 @Configuration
-@EnableConfigurationProperties(LoadbalancerConfiguration.class)
-@ConditionalOnClass(IRule.class)
-@ConfigurationProperties("tx-lcn.ribbon.loadbalancer.dtx")
-@ConditionalOnProperty(name = "tx-lcn.ribbon.loadbalancer.dtx.enabled", matchIfMissing = true)
 public class LoadbalancerConfiguration {
-    private boolean enabled;
-
-    public boolean isEnabled() {
-        return enabled;
-    }
-
-    public void setEnabled(boolean enabled) {
-        this.enabled = enabled;
-    }
 
     @Bean
-    @ConditionalOnMissingBean
+    @ConditionalOnClass(IRule.class)
+    @ConditionalOnProperty(name = "tx-lcn.ribbon.loadbalancer.dtx.enabled", havingValue = "true")
     @Scope("prototype")
     public IRule ribbonRule(Registration registration) {
         return new TxlcnZoneAvoidanceRule(registration);
